backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
supabase>=2.4.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
stripe>=10.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
resend>=2.0.0
apscheduler>=3.10.0
tenacity>=8.2.0
huggingface-hub>=0.23.0
aiohttp>=3.9.0
slowapi>=0.1.9
